# Always sanitize derivation name